made faq mobile adaptive and can scroll to top
