+++ src/components/FAQ/FAQHeader.js
@@ -3,7 +3,7 @@ import ArmadaLogo from '../../../static/assets/virtual_logo_green.svg'
 
 const FAQHeader = () => {
     return (
-        <div style={{ marginTop: '-5%' }}>
+        <div style={{ marginTop: '5%' }}>
             <header style={{ textAlign: 'center' }}>
                 <img
                     alt=''
@@ -11,7 +11,7 @@ const FAQHeader = () => {
                     style={{ width: '90px', height: '90px' }}
                 ></img>
                 <h2 id='FAQ' style={{ marginBottom: '20px' }}>
-                    How can we help you?
+                    FAQ
                 </h2>
             </header>
         </div>
